// src/interfaces.ts
export interface Options {
  testrailEnabled: boolean;
  host: string;
  user: string;
  apiKey: string;
  project: string;
  projectId: string;
  plan: string;
  planId: string;
  suite: string;
  suiteId: string;
  runName: string;
  runDescription: string;
  runCloseAfterDays: string;
  caseMeta: string;
  uploadScreenshots: boolean;
}

export interface TestResult {
  name: string;
  meta: Meta;
  caseId: number;
  testRunInfo: TestRunInfo;
  durationFormatted: string;
  testStatus: {
    value: TestStatus;
    text: string;
    color: string;
  };
}

export interface Meta {
  [key: string]: string;
}

export interface TaskResult {
  passedCount: number;
  failedCount: number;
  skippedCount: number;
}

export interface Screenshot {
  screenshotPath: string;
  thumbnailPath: string;
  userAgent: string;
  quarantineAttempt: number;
  takenOnFail: boolean;
}

export interface TestRunInfo {
  errs: any[];
  warnings: string[];
  durationMs: number;
  unstable: boolean;
  screenshotPath: string;
  screenshots: Screenshot[];
  quarantine: { [key: string]: { passed: boolean } };
  skipped: boolean;
}

export abstract class Reporter {
  newline!: () => Reporter;
  write!: (text: string) => Reporter;
  //writeLine: (line: string) => Reporter;
  useWordWrap!: (use: boolean) => Reporter;
  setIndent!: (indent: number) => Reporter;

  indentString!: (text: string, indent: number) => string;
  wordWrap!: (text: string, indent: number, width: number) => string;
  escapeHtml!: (html: string) => string;
  formatError!: (error: object, prefix?: string) => string;

  chalk!: typeof chalk;
  moment!: typeof moment;
  symbols!: {
    err: string;
  };
  abstract noColors: boolean;

  abstract async reportTaskStart(
    startTime: number,
    userAgents: string[],
    testCount: number
  ): Promise<void>;
  abstract reportFixtureStart(name: string, path: string, meta: Meta): void;
  reportTestStart?(name: string, meta: Meta): void;
  abstract reportTestDone(
    name: string,
    testRunInfo: TestRunInfo,
    meta: Meta
  ): void;
  abstract reportTaskDone(
    endTime: number,
    passed: number,
    warnings: string[],
    result: TaskResult
  ): void;
}

// Testrail interfaces
export enum TestStatus {
  Passed = 1,
  Blocked,
  Untested, // Cannot add this status when posting a result
  Retest,
  Failed,
}